requirements.txt
torch
transformers
Pillow
opencv-python
numpy
scipy
openpyxl
pandas
matplotlib